refactor(common): move hash functions to utils directory
