+++ apps/mobile/app/(auth)/signup.tsx
@@ -9,7 +9,7 @@ import { Checkbox } from "~/components/ui/checkbox";
 import React, { useState } from "react";
 import api from "@common/api/api";
 import type { RegisterDto } from "@common/types/auth/dto/register.dto";
-import hashPassword from "@common/hash/hashPassword";
+import hashPassword from "@common/utils/hash/hashPassword";
 
 export default function SignupPage() {
     const { t } = useTranslation();
