+++ packages/common/src/utils/hash/hashString.ts
